- Made some changes the loading/initialization to use load_nerf_post() and did the initialization of the Testbed function. - Changed the input of the sampleIdx and finalDepth to be normalized between 0 and 1.

+++ src/ext/instant-ngp/include/neural-graphics-primitives/nerf_device.cuh
@@ -51,8 +51,8 @@ struct TrainingImageMetadata {
 	const float* depth = nullptr;
 	const Ray* rays = nullptr;
 	// const int* pixel_indices = nullptr;
-	const int* sample_indices = nullptr;
-	const int* final_depths = nullptr;
+	const float* sample_indices = nullptr;
+	const float* final_depths = nullptr;
 
 
 	Lens lens = {};
